// src/app/features/utility-configuration/mobile/models/pages.interface.ts
export interface IMobilePageDetails {
  IsWelcomePage: boolean;
  IsGlobalQuestionPage: boolean;
  IsServicePage: boolean;
  IsServiceQuestionPage: boolean;
  IsThankYouPage: boolean;
  IsTicketPage: boolean;
  IsMarketingPage: boolean;
  IsSurveyPage: boolean;
  IsLanguagePage: boolean;
  IsNoQueuePage: boolean;
  IsOffLinePage: boolean;
}